requirements.txt
python-dotenv
msgraph-sdk
PyYAML
Flask